+++ Timer.h
@@ -11,21 +11,21 @@ namespace Victoria {
       typedef std::function<void()> TCallback;
       public:
         Timer();
-        unsigned int setTimeout(unsigned short timespan, TCallback callback);
-        unsigned int setInterval(unsigned short timespan, TCallback callback);
+        unsigned int setTimeout(int timespan, TCallback callback);
+        unsigned int setInterval(int timespan, TCallback callback);
         bool clearTimeout(unsigned int id);
         bool clearInterval(unsigned int id);
         void loop();
       private:
         struct Config {
           bool repeat; // true for setInterval, otherwise setTimeout
-          unsigned short timespan; // delay or interval in milliseconds
+          int timespan; // delay or interval in milliseconds
           TCallback callback; // callback function
           unsigned long timestamp; // timestamp when timer registered
         };
         unsigned int _idSeed = 0;
         std::map<unsigned int, Config> _configs;
-        unsigned int _addConfig(bool repeat, unsigned short timespan, TCallback callback);
+        unsigned int _addConfig(bool repeat, int timespan, TCallback callback);
         bool _removeConfig(unsigned int id);
     };
   }
